added phone number field in react library
