+++ packages/react/package.json
@@ -35,7 +35,8 @@
     "@corbado/shared-ui": "^1.1.4",
     "i18next": "23.5.1",
     "i18next-browser-languagedetector": "7.1.0",
-    "react-i18next": "13.2.2"
+    "react-i18next": "13.2.2",
+    "react-phone-number-input": "^3.3.9"
   },
   "devDependencies": {
     "@corbado/types": "^1.1.3",
