+++ packages/shared-ui/src/styles/variables.css
@@ -9,7 +9,9 @@
   --cb-dark: #000000;
   --cb-border-color: #c7c7c7;
   --cb-secondary-link-color: #59acff;
-  --cb-edit-box-color: #fafafa;
+  --cb-box-color: #fafafa;
+  --cb-box-color-hover: #c5c2c2;
+  --cb-error-text-color: #ce1212;
   --cb-primary-font: 'Space Grotesk', sans-serif;
   --cb-secondary-font: 'Inter', sans-serif;
   --cb-font-weight-bold: 700;
@@ -17,13 +19,13 @@
   --cb-base-font-size: 0.65rem;
   --cb-border-radius-lg: 1.125rem;
   --cb-border-radius-sm: 0.5rem;
-  --cb-error-text-color: #ce1212;
   --cb-error-font-size: 0.65rem;
+  --cb-input-field-height: 1.75rem;
 }
 
 .cb-dark {
   --cb-text-primary-color: #ffffff;
-  --cb-edit-box-color: #1f1f1f;
+  --cb-box-color: #1f1f1f;
   --cb-dark: #ffffff;
   --cb-white: #000000;
 }
