+++ playground/react/src/components/AuthButtons.tsx
@@ -1,13 +1,8 @@
 import { useNavigate, useParams } from 'react-router-dom';
-import 'react-phone-number-input/style.css';
-import { Value, formatPhoneNumberIntl } from 'react-phone-number-input/input';
-import { useState } from 'react';
-import CountrySelect from './CountryDropdown';
 
 export const AuthButtons = () => {
   const navigate = useNavigate();
   const { projectId } = useParams();
-  const [value, setValue] = useState<Value>();
 
   return (
     <div className='auth-buttons'>
@@ -30,12 +25,6 @@ export const AuthButtons = () => {
       >
         Login Page
       </button>
-      <CountrySelect onChange={setValue}></CountrySelect>
-      {value && (
-        <div>
-          Phone number is: {value} <p>International Format is: {formatPhoneNumberIntl(value)}</p>
-        </div>
-      )}
     </div>
   );
 };
