--- playground/react/src/components/temp.css
@@ -1,78 +1,0 @@
-.cb-phone-input-field {
-    display: flex;
-    flex-direction: row;
-    gap: 0.5rem;
-    font-family: 'Inter', sans-serif;
-    font-size: 13px;
-    height: 32px;
-    width: 315px;
-    border: 1px solid #c7c7c7;
-    border-radius: 0.5rem;
-    position: relative;
-}
-
-.cb-phone-input-field-button {
-    flex: 1;
-    background-color: #f5f5f5;
-    cursor: pointer;
-    display: flex;
-    flex-direction: row;
-    gap: 0.5rem;
-    align-items: center;
-    border: 0px;
-    padding-left: 0.8rem;
-    border-top-left-radius: 0.5rem;
-    border-bottom-left-radius: 0.5rem;
-}
-
-.cb-phone-input-field-button svg {
-    width: 20px;
-}
-
-.cb-phone-input-field-selection {
-    position: absolute;
-    width: 100%;
-    box-shadow: 0px 8px 16px 0px rgba(0, 0, 0, 0.2);
-    z-index: 1;
-    background-color: #f5f5f5;
-    top: 35px;
-    height: 250px;
-    overflow-y: auto;
-}
-
-.cb-phone-input-field-selection svg {
-    width: 20px;
-}
-
-.cb-phone-input-field-selection-item {
-    padding: 0.75rem 1.5rem;
-    cursor: pointer;
-    display: flex;
-    align-items: center;
-    gap: 0.75rem;
-    border-bottom: 1px solid #c7c7c7;
-}
-
-.cb-phone-input-field-selection-item-selected {
-    background-color: #c5c2c2;
-    cursor: none;
-    color: #1953ff;
-}
-
-.cb-phone-input-field-selection-item:hover {
-    background-color: #c5c2c2;
-}
-
-.cb-phone-input-field-input {
-    flex: 3;
-    border: 0px;
-    font-family: 'Inter', sans-serif;
-    font-size: 13px;
-    height: 95%;
-    border-top-right-radius: 0.5rem;
-    border-bottom-right-radius: 0.5rem;
-}
-
-.cb-phone-input-field-input:focus {
-    outline: none;
-}
